refactored address lib into smaller file by category, by type
