+++ packages/jellyfish-address/__tests__/P2WSH.test.ts
@@ -1,6 +1,6 @@
 import { MainNet, RegTest, TestNet } from '@defichain/jellyfish-network'
 import { OP_CODES } from '@defichain/jellyfish-transaction/src/script'
-import { DeFiAddress, P2WSH } from '../src'
+import DeFiAddress, { P2WSH } from '../src'
 
 describe('P2WSH', () => {
   const p2wshFixture = {
